Refactor out WKSourceEditorView to simplify communication
