+++ Components/Sources/Components/Components/Editors/Source Editor/WKSourceEditorViewController.swift
@@ -5,15 +5,150 @@ public protocol WKSourceEditorViewControllerDelegate: AnyObject {
     func sourceEditorViewControllerDidTapFind(sourceEditorViewController: WKSourceEditorViewController)
 }
 
+// MARK: NSNotification Names
+
+extension Notification.Name {
+    static let WKSourceEditorSelectionState = Notification.Name("WKSourceEditorSelectionState")
+}
+
+extension Notification {
+    static let WKSourceEditorSelectionState = Notification.Name.WKSourceEditorSelectionState
+    static let WKSourceEditorSelectionStateKey = "WKSourceEditorSelectionStateKey"
+}
+
 public class WKSourceEditorViewController: WKComponentViewController {
+    
+    // MARK: Nested Types
+    
+    enum InputViewType {
+        case main
+        case headerSelect
+    }
+    
+    enum InputAccessoryViewType {
+        case expanding
+        case highlight
+        case find
+    }
     
     // MARK: - Properties
     
     private let viewModel: WKSourceEditorViewModel
     private weak var delegate: WKSourceEditorViewControllerDelegate?
+    private let textFrameworkMediator: WKSourceEditorTextFrameworkMediator
     
-    private var editorView: WKSourceEditorView {
-        return view as! WKSourceEditorView
+    var textView: UITextView {
+        return textFrameworkMediator.textView
+    }
+    
+    // Input Accessory Views
+    
+    private lazy var expandingAccessoryView: WKEditorToolbarExpandingView = {
+        let view = UINib(nibName: String(describing: WKEditorToolbarExpandingView.self), bundle: Bundle.module).instantiate(withOwner: nil).first as! WKEditorToolbarExpandingView
+        view.delegate = self
+        return view
+    }()
+    
+    private lazy var highlightAccessoryView: WKEditorToolbarHighlightView = {
+        let view = UINib(nibName: String(describing: WKEditorToolbarHighlightView.self), bundle: Bundle.module).instantiate(withOwner: nil).first as! WKEditorToolbarHighlightView
+        view.delegate = self
+        
+        return view
+    }()
+    
+    private lazy var findAccessoryView: WKFindAndReplaceView = {
+        let view = UINib(nibName: String(describing: WKFindAndReplaceView.self), bundle: Bundle.module).instantiate(withOwner: nil).first as! WKFindAndReplaceView
+        let viewModel = WKFindAndReplaceViewModel()
+        view.configure(viewModel: viewModel)
+        
+        return view
+    }()
+    
+    // Input Views
+    
+    private var _mainInputView: UIView?
+    private var mainInputView: UIView? {
+        get {
+            guard _mainInputView == nil else {
+                return _mainInputView
+            }
+            
+            let inputViewController = WKEditorInputViewController(configuration: .rootMain, delegate: self)
+            inputViewController.loadViewIfNeeded()
+            
+            _mainInputView = inputViewController.view
+            
+            return inputViewController.view
+        }
+        set {
+            _mainInputView = newValue
+        }
+    }
+    
+    private var _headerSelectionInputView: UIView?
+    private var headerSelectionInputView: UIView? {
+        get {
+            guard _headerSelectionInputView == nil else {
+                return _headerSelectionInputView
+            }
+            
+            let inputViewController = WKEditorInputViewController(configuration: .rootHeaderSelect, delegate: self)
+            inputViewController.loadViewIfNeeded()
+            
+            _headerSelectionInputView = inputViewController.view
+            
+            return inputViewController.view
+        }
+        set {
+            _headerSelectionInputView = newValue
+        }
+    }
+    
+    // Input Types
+    
+    var inputViewType: InputViewType? = nil {
+        didSet {
+            
+            guard let inputViewType else {
+                mainInputView = nil
+                headerSelectionInputView = nil
+                textView.inputView = nil
+                textView.reloadInputViews()
+                return
+            }
+            
+            switch inputViewType {
+            case .main:
+                textView.inputView = mainInputView
+            case .headerSelect:
+                textView.inputView = headerSelectionInputView
+            }
+            
+            textView.inputAccessoryView = nil
+            textView.reloadInputViews()
+        }
+    }
+    var inputAccessoryViewType: InputAccessoryViewType? = nil {
+        didSet {
+            
+            guard let inputAccessoryViewType else {
+                textView.inputAccessoryView = nil
+                textView.reloadInputViews()
+                return
+            }
+            
+            switch inputAccessoryViewType {
+            case .expanding:
+                textView.inputAccessoryView = expandingAccessoryView
+            case .highlight:
+                textView.inputAccessoryView = highlightAccessoryView
+            case .find:
+                textView.inputAccessoryView = findAccessoryView
+            }
+            
+            textView.inputView = nil
+            textView.reloadInputViews()
+        }
     }
     
     // MARK: - Lifecycle
@@ -21,40 +156,107 @@ public class WKSourceEditorViewController: WKComponentViewController {
     public init(viewModel: WKSourceEditorViewModel, delegate: WKSourceEditorViewControllerDelegate) {
         self.viewModel = viewModel
         self.delegate = delegate
+        self.textFrameworkMediator = WKSourceEditorTextFrameworkMediator()
         super.init()
+        setup()
     }
-
+    
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    public override func loadView() {
-        self.view = WKSourceEditorView(delegate: self)
+    private func setup() {
+        textView.delegate = self
+        view.addSubview(textView)
+        updateColorsAndFonts()
+        
+        NSLayoutConstraint.activate([
+            view.safeAreaLayoutGuide.leadingAnchor.constraint(equalTo: textView.leadingAnchor),
+            view.safeAreaLayoutGuide.trailingAnchor.constraint(equalTo: textView.trailingAnchor),
+            view.safeAreaLayoutGuide.topAnchor.constraint(equalTo: textView.topAnchor),
+            view.safeAreaLayoutGuide.bottomAnchor.constraint(equalTo: textView.bottomAnchor)
+        ])
+        
+        NotificationCenter.default.addObserver(self,
+                                               selector: #selector(keyboardWillChangeFrame(_:)),
+                                               name: UIApplication.keyboardWillChangeFrameNotification,
+                                               object: nil)
+        NotificationCenter.default.addObserver(self,
+                                               selector: #selector(keyboardWillHide(_:)),
+                                               name: UIApplication.keyboardWillHideNotification,
+                                               object: nil)
     }
     
     public override func viewDidLoad() {
         super.viewDidLoad()
-
-        editorView.setup(viewModel: viewModel)
-        editorView.inputAccessoryViewType = .expanding
+        
+        setup(viewModel: viewModel)
+        inputAccessoryViewType = .expanding
+    }
+    
+    // MARK: Overrides
+    
+    public override func appEnvironmentDidChange() {
+        updateColorsAndFonts()
+    }
+    
+    // MARK: - Notifications
+    
+    @objc private func keyboardWillHide(_ notification: Notification) {
+        updateInsets(keyboardHeight: 0)
+    }
+    
+    @objc private func keyboardWillChangeFrame(_ notification: Notification) {
+        if let frame = notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect {
+            let height = max(frame.height - view.safeAreaInsets.bottom, 0)
+            updateInsets(keyboardHeight: height)
+        }
     }
     
     // MARK: - Public
     
     public func closeFind() {
-        editorView.closeFind()
-        editorView.inputAccessoryViewType = .expanding
+        textView.becomeFirstResponder()
+        inputAccessoryViewType = .expanding
     }
     
     public func toggleSyntaxHighlighting() {
         viewModel.isSyntaxHighlightingEnabled.toggle()
-        editorView.update(viewModel: viewModel)
+        update(viewModel: viewModel)
+    }
+}
+
+// MARK: - Private
+
+private extension WKSourceEditorViewController {
+
+    func setup(viewModel: WKSourceEditorViewModel) {
+        textFrameworkMediator.isSyntaxHighlightingEnabled = viewModel.isSyntaxHighlightingEnabled
+        textView.attributedText = NSAttributedString(string: viewModel.initialText)
     }
     
-    // MARK: - Private
+    func update(viewModel: WKSourceEditorViewModel) {
+        textFrameworkMediator.isSyntaxHighlightingEnabled = viewModel.isSyntaxHighlightingEnabled
+    }
     
-    private func postUpdateButtonSelectionStatesNotification(withDelay delay: Bool) {
-        let selectionState = editorView.selectionState()
+    func updateInsets(keyboardHeight: CGFloat) {
+        textView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: keyboardHeight, right: 0)
+        textView.scrollIndicatorInsets = UIEdgeInsets(top: 0, left: 0, bottom: keyboardHeight, right: 0)
+    }
+    
+    func updateColorsAndFonts() {
+        view.backgroundColor = WKAppEnvironment.current.theme.paperBackground
+        textView.backgroundColor = WKAppEnvironment.current.theme.paperBackground
+        textView.keyboardAppearance = WKAppEnvironment.current.theme.keyboardAppearance
+        textFrameworkMediator.updateColorsAndFonts()
+    }
+    
+    func selectionState() -> WKSourceEditorSelectionState {
+        return textFrameworkMediator.selectionState(selectedDocumentRange: textView.selectedRange)
+    }
+    
+    func postUpdateButtonSelectionStatesNotification(withDelay delay: Bool) {
+        let selectionState = selectionState()
         if delay {
             DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
                 NotificationCenter.default.post(name: Notification.WKSourceEditorSelectionState, object: nil, userInfo: [Notification.WKSourceEditorSelectionStateKey: selectionState])
@@ -65,61 +267,77 @@ public class WKSourceEditorViewController: WKComponentViewController {
     }
 }
 
-// MARK: - WKSourceEditorViewDelegate
+// MARK: - UITextViewDelegate
 
-extension WKSourceEditorViewController: WKSourceEditorViewDelegate {
-    func editorViewDidTapItalics(editorView: WKSourceEditorView, isSelected: Bool) {
-        let action: WKSourceEditorFormatterButtonAction = isSelected ? .remove : .add
-        editorView.toggleItalicsFormatting(action: action, in: editorView.textView)
-    }
-    
-    func editorViewDidTapBold(editorView: WKSourceEditorView, isSelected: Bool) {
-        let action: WKSourceEditorFormatterButtonAction = isSelected ? .remove : .add
-        editorView.toggleBoldFormatting(action: action, in: editorView.textView)
-    }
-    
-    func editorViewTextSelectionDidChange(editorView: WKSourceEditorView, isRangeSelected: Bool) {
-        guard editorView.inputViewType == nil else {
+extension WKSourceEditorViewController: UITextViewDelegate {
+    public func textViewDidChangeSelection(_ textView: UITextView) {
+        guard inputViewType == nil else {
             postUpdateButtonSelectionStatesNotification(withDelay: false)
             return
         }
-        
-        editorView.inputAccessoryViewType = isRangeSelected ? .highlight : .expanding
+        let isRangeSelected = textView.selectedRange.length > 0
+        inputAccessoryViewType = isRangeSelected ? .highlight : .expanding
         postUpdateButtonSelectionStatesNotification(withDelay: false)
     }
-    
-    func editorViewDidTapFind(editorView: WKSourceEditorView) {
-        editorView.inputAccessoryViewType = .find
+}
+
+// MARK: - WKEditorToolbarExpandingViewDelegate
+
+extension WKSourceEditorViewController: WKEditorToolbarExpandingViewDelegate {
+    func toolbarExpandingViewDidTapFind(toolbarView: WKEditorToolbarExpandingView) {
+        inputAccessoryViewType = .find
         delegate?.sourceEditorViewControllerDidTapFind(sourceEditorViewController: self)
     }
     
-    func editorViewDidTapFormatText(editorView: WKSourceEditorView) {
-        editorView.inputViewType = .main
+    func toolbarExpandingViewDidTapFormatText(toolbarView: WKEditorToolbarExpandingView) {
+        inputViewType = .main
         postUpdateButtonSelectionStatesNotification(withDelay: true)
     }
     
-    func editorViewDidTapFormatHeading(editorView: WKSourceEditorView) {
-        editorView.inputViewType = .headerSelect
+    func toolbarExpandingViewDidTapFormatHeading(toolbarView: WKEditorToolbarExpandingView) {
+        inputViewType = .headerSelect
+    }
+}
+
+// MARK: - WKEditorToolbarHighlightViewDelegate
+
+extension WKSourceEditorViewController: WKEditorToolbarHighlightViewDelegate {
+    func toolbarHighlightViewDidTapBold(toolbarView: WKEditorToolbarHighlightView, isSelected: Bool) {
+        let action: WKSourceEditorFormatterButtonAction = isSelected ? .remove : .add
+        textFrameworkMediator.boldItalicsFormatter?.toggleBoldFormatting(action: action, in: textView)
     }
     
-    func editorViewDidTapCloseInputView(editorView: WKSourceEditorView, isRangeSelected: Bool) {
-        editorView.inputViewType = nil
-        editorView.inputAccessoryViewType = isRangeSelected ? .highlight : .expanding
+    func toolbarHighlightViewDidTapItalics(toolbarView: WKEditorToolbarHighlightView, isSelected: Bool) {
+        let action: WKSourceEditorFormatterButtonAction = isSelected ? .remove : .add
+        textFrameworkMediator.boldItalicsFormatter?.toggleItalicsFormatting(action: action, in: textView)
     }
     
-    func editorViewDidTapShowMore(editorView: WKSourceEditorView) {
-        editorView.inputViewType = .main
+    func toolbarHighlightViewDidTapShowMore(toolbarView: WKEditorToolbarHighlightView) {
+        inputViewType = .main
         postUpdateButtonSelectionStatesNotification(withDelay: true)
     }
+    
+    func toolbarHighlightViewDidTapFormatHeading(toolbarView: WKEditorToolbarHighlightView) {
+        inputViewType = .headerSelect
+    }
 }
 
-// MARK: NSNotification Names
+// MARK: - WKEditorInputViewDelegate
 
-extension Notification.Name {
-    static let WKSourceEditorSelectionState = Notification.Name("WKSourceEditorSelectionState")
-}
-
-extension Notification {
-    static let WKSourceEditorSelectionState = Notification.Name.WKSourceEditorSelectionState
-    static let WKSourceEditorSelectionStateKey = "WKSourceEditorSelectionStateKey"
+extension WKSourceEditorViewController: WKEditorInputViewDelegate {
+    func didTapBold(isSelected: Bool) {
+        let action: WKSourceEditorFormatterButtonAction = isSelected ? .remove : .add
+        textFrameworkMediator.boldItalicsFormatter?.toggleBoldFormatting(action: action, in: textView)
+    }
+    
+    func didTapItalics(isSelected: Bool) {
+        let action: WKSourceEditorFormatterButtonAction = isSelected ? .remove : .add
+        textFrameworkMediator.boldItalicsFormatter?.toggleItalicsFormatting(action: action, in: textView)
+    }
+    
+    func didTapClose() {
+        inputViewType = nil
+        let isRangeSelected = textView.selectedRange.length > 0
+        inputAccessoryViewType = isRangeSelected ? .highlight : .expanding
+    }
 }
